request context abstraction, refactoring
