+++ src/main/java/org/folio/rest/migration/model/request/AbstractJob.java
@@ -1,0 +1,33 @@
+package org.folio.rest.migration.model.request;
+
+import javax.validation.constraints.NotNull;
+
+public abstract class AbstractJob {
+
+  @NotNull
+  private String schema;
+
+  @NotNull
+  private int partitions;
+
+  public AbstractJob() {
+
+  }
+
+  public String getSchema() {
+    return schema;
+  }
+
+  public void setSchema(String schema) {
+    this.schema = schema;
+  }
+
+  public int getPartitions() {
+    return partitions;
+  }
+
+  public void setPartitions(int partitions) {
+    this.partitions = partitions;
+  }
+
+}
