--- src/main/java/org/folio/rest/migration/model/request/Extraction.java
@@ -1,57 +1,0 @@
-package org.folio.rest.migration.model.request;
-
-import javax.validation.constraints.NotNull;
-
-import org.folio.rest.migration.config.model.Database;
-
-public class Extraction {
-
-  @NotNull
-  private String count;
-
-  @NotNull
-  private String page;
-
-  @NotNull
-  private String additional;
-
-  @NotNull
-  private Database database;
-
-  public Extraction() {
-
-  }
-
-  public String getCount() {
-    return count;
-  }
-
-  public void setCount(String count) {
-    this.count = count;
-  }
-
-  public String getPage() {
-    return page;
-  }
-
-  public void setPage(String page) {
-    this.page = page;
-  }
-
-  public String getAdditional() {
-    return additional;
-  }
-
-  public void setAdditional(String additional) {
-    this.additional = additional;
-  }
-
-  public Database getDatabase() {
-    return database;
-  }
-
-  public void setDatabase(Database database) {
-    this.database = database;
-  }
-
-}
